app/requirements.txt
bcrypt==3.1.7
email-validator==1.0.5
fastapi==0.49.0
gino==0.8.3
PyJWT==1.7.1
uvicorn==0.11.3
passlib==1.7.1